% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidphdata.R
\docType{package}
\name{covidphdata}
\alias{covidphdata}
\alias{covidphdata-package}
\title{An Interface to the Philippines Department of Health COVID-19 Data Drop}
\description{
The Philippines \href{https://www.doh.gov.ph}{Department of Health} has made
\emph{COVID-19} related data publicly available as part of its mandate to
promoting transparency and accountability in governance. The Philippines
\strong{COVID-19 Data Drop} is distributed via \emph{Google Drive} with latest updated
data provided daily. Data from previous days are archived and also made
available through \emph{Google Drive}. This package provides a coherent and
performant API to the latest and archived Philippines \emph{COVID-19} data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://como-ph.github.io/covidphdata/}
  \item Report bugs at \url{https://github.com/como-ph/covidphdata/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
