% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datadrop_download}
\alias{datadrop_download}
\title{Get link to a Google Drive file and download}
\usage{
datadrop_download(id, path)
}
\arguments{
\item{id}{A 33-character string identifier for the Google Drive file}

\item{path}{Path to download the file to.}
}
\value{
A file downloaded into specified path
}
\description{
Get link to a Google Drive file and download
}
\author{
Ernest Guevarra
}
