% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{datadrop_fields}
\alias{datadrop_fields}
\title{Pull Philippines data fields information from the publicly available
Department of Health COVID-19 Data Drop}
\usage{
datadrop_fields(version = "current", .date = NULL)
}
\arguments{
\item{version}{A character value specifying whether to get the most
currently available dataset (\code{"current"}) or to get archive data
(\code{"archive"}). Default to \code{"current"}}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date up to which extracted data reports to. Only used when \code{version} is set to
\code{"archive"} otherwise ignored.}
}
\value{
A tibble of metadata for the fields used in the various datasets
distributed via the COVID-19 Data Drop
}
\description{
A wrapper to googledrive and googlesheets4 functions to pull data from the
Philppines' COVID-19 Data Drop resource that is publicly distributed via
Google Drive
}
\examples{
datadrop_fields()

}
