% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{datadrop_changelog}
\alias{datadrop_changelog}
\alias{datadrop_sheets}
\alias{datadrop_fields}
\alias{datadrop_cases}
\alias{datadrop_tests}
\alias{datadrop_collect_daily}
\alias{datadrop_collect_weekly}
\alias{datadrop_quarantine_daily}
\alias{datadrop_quarantine_weekly}
\title{Retrieve datasets from specified DoH Data Drop folders}
\usage{
datadrop_changelog(version = c("latest", "archive"), .date = NULL)

datadrop_sheets(version = c("latest", "archive"), .date = NULL)

datadrop_fields(version = c("latest", "archive"), .date = NULL)

datadrop_cases(version = c("latest", "archive"), .date = NULL)

datadrop_tests(version = c("latest", "archive"), .date = NULL)

datadrop_collect_daily(version = c("latest", "archive"), .date = NULL)

datadrop_collect_weekly(version = c("latest", "archive"), .date = NULL)

datadrop_quarantine_daily(version = c("latest", "archive"), .date = NULL)

datadrop_quarantine_weekly(version = c("latest", "archive"), .date = NULL)
}
\arguments{
\item{version}{A character value specifying whether to get the latest
available dataset (\code{latest}) or to get archive data (\code{archive}). Default
to \code{latest}.}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date for the archive DoH Data Drop for which an ID is to be returned.
Should be specified when using \code{version} is set to \code{archive} otherwise
ignored.}
}
\value{
A tibble of any of the following datasets:
\itemize{
\item 
\enumerate{
\item \strong{Metadata - Sheets};
}
\item 
\enumerate{
\item \strong{Metadata - Fields};
}
\item 
\enumerate{
\item \strong{Case Information};
}
\item 
\enumerate{
\item \strong{DOH Data Collect - Daily Report};
}
\item 
\enumerate{
\item \strong{DOH Data Collect - Weekly Report};
}
\item 
\enumerate{
\item \strong{Testing Aggregates};
}
\item 
\enumerate{
\item \strong{Quarantine Facility Data - Daily Report};
}
\item 
\enumerate{
\item \strong{Quarantine Facility Data - Weekly Report};
}
\item 
\enumerate{
\item \strong{DOH Data Collect v3 - Baseline};
}
\item 
\enumerate{
\item \strong{DOH Data Collect v4 - Baseline}; and,
}
\item 
\enumerate{
\item \strong{DDC TTMF Tracker v1}.
For Changelog, a named list of two tibbles - \emph{List of Changes} and
\emph{Most Common Changes}.
}
}
}
\description{
A wrapper to \code{googledrive} functions to retrieve datasets from the DoH Data
Drop folders
}
\examples{
datadrop_changelog()
datadrop_sheets()
datadrop_fields()
datadrop_cases()
datadrop_collect_daily()
datadrop_collect_weekly()
datadrop_tests()
datadrop_quarantine_daily()
datadrop_quarantine_weekly()
#datadrop_collect_v3()
#datadrop_collect_v4()
#datadrop_tracker()

}
