% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datadrop_ls}
\alias{datadrop_ls}
\title{List files and folders within a specified DoH Data Drop Google Drive folder}
\usage{
datadrop_ls(id)
}
\arguments{
\item{id}{A 33 character string for the \emph{Google Drive} ID of the latest
officially released \strong{DoH Data Drop}. This can be obtained using the
\code{\link[=datadrop_id_latest]{datadrop_id_latest()}} function or by manually expanding the shortened URL
provided in \href{https://bit.ly/DataDropPH}{bit.ly/DataDropPH} and extracting the
\emph{Google Drive} ID.}
}
\value{
A tibble containing information on the various files and folders in
the specified \emph{Google Drive} for \strong{DoH Data Drop}.
}
\description{
List files and folders within a specified DoH Data Drop Google Drive folder
}
\examples{
\dontrun{
  library(googledrive)

  ## Authentication
  googledrive::drive_auth_configure(api_key = Sys.getenv("GOOGLEDRIVE_TOKEN"))

  ## Deauthorise
  googledrive::drive_deauth()

  ## Get Google Drive ID of latest DoH Data Drop
  id <- datadrop_id_latest()

  ## List the contents of the latest DoH Data Drop in Google Drive
  datadrop_ls(id = id)
}

}
\author{
Ernest Guevarra
}
