% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_data.R
\name{datadrop_ls_all}
\alias{datadrop_ls_all}
\title{Get Philippines Department of Health COVID-19 Data Drop Google Drive specific
directory information}
\usage{
datadrop_ls_all(version = "current", .date = NULL)
}
\arguments{
\item{version}{A character value specifying whether to get the most
currently available dataset (\code{"current"}) or to get archive data
(\code{"archive"}). Default to \code{"current"}}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date up to which extracted data reports to. Only used when \code{version} is set to
\code{"archive"} otherwise ignored.}
}
\value{
A tibble containing information on the various files in a specified
\strong{COVID-19 Data Drop Google Drive} directory
}
\description{
A wrapper to \href{https://googledrive.tidyverse.org}{googledrive package}
functions to extract information on contents of a specific
\strong{COVID-19 Data Drop Google Drive} directory
}
\examples{
datadrop_ls_all()

}
