% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datadrop_download}
\alias{datadrop_download}
\title{Download DoH Data Drop file via its Google Drive ID}
\usage{
datadrop_download(id, path = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{id}{A 33-character string identifier for the \emph{Google Drive} file.}

\item{path}{A character value for path for output file. If NULL, the
default file name used in \emph{Google Drive} is used and the default location is
the working directory.}

\item{overwrite}{Logical. If \code{path} already exists, should it be overwritten?
Default to FALSE.}

\item{verbose}{Logical. Should operation progress messages be shown? Default
to TRUE.}
}
\value{
A file downloaded into specified path
}
\description{
Download DoH Data Drop file via its Google Drive ID
}
\examples{
## Get Google Drive ID for Case Information file in latest DoH Data Drop
id <- datadrop_id_file(tbl = datadrop_ls(id = datadrop_id()),
                       fn = "Metadata - Sheets.csv")

## Download the Case Information file into tempfile()
datadrop_download(id = id, path = tempfile())

}
\author{
Ernest Guevarra
}
