% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{datadrop_get}
\alias{datadrop_get}
\alias{datadrop_get_changelog}
\alias{datadrop_get_sheets}
\alias{datadrop_get_fields}
\alias{datadrop_get_cases}
\alias{datadrop_get_tests}
\alias{datadrop_get_cdaily}
\alias{datadrop_get_cweekly}
\alias{datadrop_get_qdaily}
\alias{datadrop_get_qweekly}
\title{Retrieve datasets from specified DoH Data Drop folders}
\usage{
datadrop_get(id, path = NULL, keep = FALSE, overwrite = FALSE, verbose = TRUE)

datadrop_get_changelog(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{A 33-character string identifier for the \emph{Google Drive} file}

\item{path}{A character value for path for output file. If NULL, the
default file name used in \emph{Google Drive} is used and the default location
is the working directory.}

\item{keep}{Logical. Should file be saved locally? Default to FALSE. If TRUE,
file is kept in the location specified in \code{path}. If \code{path} is NULL, the
file is kept in the working directory using the default file name used in
\emph{Google Drive}.}

\item{overwrite}{Logical. If \code{path} already exists, should it be overwritten?
Default to FALSE.}

\item{verbose}{Logical. Should operation progress messages be shown? Default
to TRUE.}

\item{version}{A character value specifying whether to get the latest
available dataset (\code{latest}) or to get archive data (\code{archive}). Default
to \code{latest}.}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date for the archive \strong{DoH Data Drop} for which an ID is to be returned.
Should be specified when using \code{version} is set to \code{archive} otherwise
ignored.}
}
\value{
A tibble of retrieved \strong{DoH Data Drop} file. If \code{keep} is TRUE, a
file is also downloaded into specified \code{path}.

A tibble of any of the following datasets:
\itemize{
\item 
\enumerate{
\item \emph{Metadata - Sheets};
}
\item 
\enumerate{
\item \emph{Metadata - Fields};
}
\item 
\enumerate{
\item \emph{Case Information};
}
\item 
\enumerate{
\item \emph{DOH Data Collect - Daily Report};
}
\item 
\enumerate{
\item \emph{DOH Data Collect - Weekly Report};
}
\item 
\enumerate{
\item \emph{Testing Aggregates};
}
\item 
\enumerate{
\item \emph{Quarantine Facility Data - Daily Report};
}
\item 
\enumerate{
\item \emph{Quarantine Facility Data - Weekly Report};
}
\item 
\enumerate{
\item \emph{DOH Data Collect v3 - Baseline};
}
\item 
\enumerate{
\item \emph{DOH Data Collect v4 - Baseline}; and,
}
\item 
\enumerate{
\item \emph{DDC TTMF Tracker v1}.
For Changelog, a named list of two tibbles - \emph{List of Changes} and
\emph{Most Common Changes}. If \code{keep} is TRUE, a copy of the specified
\strong{DoH Data Drop} \emph{Google Drive} file is saved in the location specified by
\code{path}.
}
}
}
\description{
A wrapper to \code{googledrive} functions to retrieve datasets from the \strong{DoH Data
Drop} folders.
}
\examples{
## Get Google Drive ID for latest case information data
id <- datadrop_id_file(tbl = datadrop_ls(id = datadrop_id()),
                       fn = "Case")

## Retrieve case information data
datadrop_get(id = id, path = tempfile())

## Retrieve latest changelog information
datadrop_get_changelog(path = tempfile())

## Retrieve latest metadata - sheets information
datadrop_get_sheets(path = tempfile())

## Retrieve latest metadata - fields information
datadrop_get_fields(path = tempfile())

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(path = tempfile())

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(path = tempfile())

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(path = tempfile())

## Retrieve latest testing aggregates information
datadrop_get_tests(path = tempfile())

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(path = tempfile())

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(path = tempfile())

#datadrop_get_collectV3(path = tempfile())
#datadrop_get_collectV4(path = tempfile())
#datadrop_get_tracker(path = tempfile())

}
