% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{datadrop_get}
\alias{datadrop_get}
\alias{datadrop_get_changelog}
\alias{datadrop_get_sheets}
\alias{datadrop_get_fields}
\alias{datadrop_get_cases}
\alias{datadrop_get_tests}
\alias{datadrop_get_cdaily}
\alias{datadrop_get_cweekly}
\alias{datadrop_get_qdaily}
\alias{datadrop_get_qweekly}
\title{Retrieve datasets from specified DoH Data Drop folders}
\usage{
datadrop_get(
  tbl,
  fn,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_changelog(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{tbl}{A tibble output produced by \code{\link[=datadrop_ls]{datadrop_ls()}} that lists the files
within a particular \strong{DoH Data Drop} \emph{Google Drive} folder}

\item{fn}{A character string composed of a word or words that can be used to
match to the name of a file within a particular \strong{DoH Data Drop}
\emph{Google Drive} folder listed in \code{tbl}.}

\item{path}{A character value for path for output file. If NULL, the
default file name used in \emph{Google Drive} is used and the default location
is the working directory.}

\item{keep}{Logical. Should file be saved locally? Default to FALSE. If TRUE,
file is kept in the location specified in \code{path}. If \code{path} is NULL, the
file is kept in the working directory using the default file name used in
\emph{Google Drive}.}

\item{overwrite}{Logical. If \code{path} already exists, should it be overwritten?
Default to FALSE.}

\item{verbose}{Logical. Should operation progress messages be shown? Default
to TRUE.}
}
\value{
A tibble of any of the following datasets:
\enumerate{
\item \emph{Metadata - Sheets};
\item \emph{Metadata - Fields};
\item \emph{Case Information};
\item \emph{DOH Data Collect - Daily Report};
\item \emph{DOH Data Collect - Weekly Report};
\item \emph{Testing Aggregates};
\item \emph{Quarantine Facility Data - Daily Report};
\item \emph{Quarantine Facility Data - Weekly Report};
\item \emph{DOH Data Collect v3 - Baseline};
\item \emph{DOH Data Collect v4 - Baseline}; and,
\item \emph{DDC TTMF Tracker v1}.
For Changelog, a named list of two tibbles - \emph{List of Changes} and
\emph{Most Common Changes}. If \code{keep} is TRUE, a copy of the specified
\strong{DoH Data Drop} \emph{Google Drive} file is saved in the location specified by
\code{path}.
}
}
\description{
A wrapper to \code{googledrive} functions to retrieve datasets from the \strong{DoH Data
Drop} folders.
}
\examples{
## Get tbl for files in latest Data Drop
library(magrittr)
x <- datadrop_id() \%>\% datadrop_ls()

## Retrieve case information data
datadrop_get(tbl = x, fn = "Case Information", path = tempfile())

## Retrieve latest changelog information
datadrop_get_changelog(tbl = x, path = tempfile())

## Retrieve latest metadata - sheets information
datadrop_get_sheets(tbl = x, path = tempfile())

## Retrieve latest metadata - fields information
datadrop_get_fields(tbl = x, path = tempfile())

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(tbl = x, path = tempfile())

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(tbl = x, path = tempfile())

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(tbl = x, path = tempfile())

## Retrieve latest testing aggregates information
datadrop_get_tests(tbl = x, path = tempfile())

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(tbl = x, path = tempfile())

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(tbl = x, path = tempfile())

#datadrop_get_collectV3(tbl = x, path = tempfile())
#datadrop_get_collectV4(tbl = x, path = tempfile())
#datadrop_get_tracker(tbl = x, path = tempfile())

}
