% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{datadrop_id_latest}
\alias{datadrop_id_latest}
\alias{datadrop_id_archive}
\alias{datadrop_id}
\alias{datadrop_id_file}
\title{Get Google Drive ID for latest or archive DoH Data Drop folders}
\usage{
datadrop_id_latest(verbose = TRUE)

datadrop_id_archive(verbose = TRUE, .date = NULL)

datadrop_id(verbose = TRUE, version = c("latest", "archive"), .date = NULL)

datadrop_id_file(tbl, fn)
}
\arguments{
\item{verbose}{Logical. Should message on operation progress be shown.
Default is TRUE.}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date for the archive \strong{DoH Data Drop} for which an ID is to be returned.
Should be specified when using \code{\link[=datadrop_id_archive]{datadrop_id_archive()}}. For \code{\link[=datadrop_id]{datadrop_id()}},
only used when \code{version} is set to \code{archive} otherwise ignored.}

\item{version}{A character value specifying whether to get the latest
available \strong{DoH Data Drop} (\code{latest}) or to get \strong{DoH Data Drop} archive
(\code{archive}). Default to \code{latest}.}

\item{tbl}{A tibble output produced by \code{\link[=datadrop_ls]{datadrop_ls()}} that lists the files
within a particular \strong{DoH Data Drop} \emph{Google Drive} folder}

\item{fn}{A character string composed of a word or words that can be used to
match to the name of a file within a particular \strong{DoH Data Drop}
\emph{Google Drive} folder listed in \code{tbl}.}
}
\value{
A 33-character string for the \emph{Google Drive} ID of the latest
\strong{DoH Data Drop} or the archive \strong{DoH Data Drop}

A 33-character string for the \emph{Google Drive} ID of the specified
\strong{DoH Data Drop} file. If \code{fn} matches with more than one file, a vector of
33-character strings for the Google Drive IDs of the specified DoH Data
Drop files.
}
\description{
The \strong{DoH Data Drop} is distributed using \emph{Google Drive} with the latest
data released through a new \emph{Google Drive} folder and the older data archived
into the same persistent \emph{Google Drive} folder.
}
\details{
The Philippines Department of Health (DoH) currently distributes the latest
\strong{Data Drop} via a fixed shortened URL (\href{https://bit.ly/DataDropPH}{bit.ly/DataDropPH})
which links/points to a new \emph{Google Drive} endpoint daily or whenever the
daily updated data drop is available. This \emph{Google Drive} endpoint is a
\strong{README} document in portable document format (PDF) which contains a
privacy and confidentiality statement, technical notes with regard to the
latest data, technical notes with regard to previous (archive data) and two
shortened URLs - one linking to the \emph{Google Drive} folder that contains all
the latest officially released datasets, and the other linking to the
datasets released previously (archives). Of these, the first shortened URL
linking to the \emph{Google Drive} folder containing the latest officially released
datasets is different for every release and can only be obtained through the
\strong{README} document released for a specific day.

The function \code{\link[=datadrop_id_latest]{datadrop_id_latest()}} reads the \strong{README} PDF file, extracts the
shortened URL for the latest official released datasets written in that file,
expands that shortened URL and then extracts the unique \emph{Google Drive} ID for
the latest officially released datasets. With this \emph{Google Drive} ID, other
functions can then be used to retrieve information and data from the Google
Drive specified by this ID.

The \strong{DoH Data Drop} archives, on the other hand, is distributed via a fixed
shortened URL (\href{https://bit.ly/DataDropArchives}{bit.ly/DataDropArchives})
which links/points to a \emph{Google Drive} folder containing the previous
\strong{DoH Data Drop} releases.

The function \code{\link[=datadrop_id_archive]{datadrop_id_archive()}} expands that shortened URL and then
extracts the unique \emph{Google Drive} ID for the \strong{DoH Data Drop} archives folder.
With this \emph{Google Drive} ID, other functions can then be used to retrieve
information and data from the \emph{Google Drive} specified by this ID.
}
\examples{
## Two ways to get the Google Drive ID of the latest DoH Data Drop
datadrop_id_latest()
datadrop_id()

## Two ways to get the Google Drive ID of the archive DoH Data Drop for
## 1 November 2020
datadrop_id_archive(.date = "2020-11-01")
datadrop_id(version = "archive", .date = "2020-11-01")

## Typical workflow
tbl <- datadrop_ls(id = datadrop_id())
datadrop_id_file(tbl = tbl, fn = "Case Information")

## Piped workflow using magrittr \%>\%
library(magrittr)

## Get the id for the latest Case Information file
datadrop_id() \%>\%
  datadrop_ls() \%>\%
  datadrop_id_file(fn = "Case Information")

}
\author{
Ernest Guevarra

Ernest Guevarra
}
