% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{datadrop_latest_id}
\alias{datadrop_latest_id}
\alias{datadrop_archive_id}
\alias{datadrop_id}
\title{Get Google Drive IDs for latest or archive DoH Data Drop folders}
\usage{
datadrop_latest_id(verbose = TRUE)

datadrop_archive_id(verbose = TRUE, .date = NULL)

datadrop_id(verbose = TRUE, version = c("latest", "archive"), .date = NULL)
}
\arguments{
\item{verbose}{Logical. Should message on operation progress be shown.
Default is TRUE.}

\item{.date}{A character value for date in \emph{YYYY-MM-DD} format. This is the
date for the archive DoH Data Drop for which an ID is to be returned.
Should be specified when using \code{\link[=datadrop_archive_id]{datadrop_archive_id()}}. For \code{\link[=datadrop_id]{datadrop_id()}},
only used when \code{version} is set to \code{archive} otherwise ignored.}

\item{version}{A character value specifying whether to get the latest
available DoH Data Drop (\code{latest}) or to get DoH Data Drop archive
(\code{archive}). Default to \code{latest}.}
}
\value{
A 33-character string for the Google Drive ID of the latest
DoH Data Drop or the archive DoH Data Drop
}
\description{
The Philippines Department of Health currently distributes the latest data
drop via a fixed shortened URL (\url{bit.ly/DataDropPH})
which links/points to a new Google Drive endpoint daily or whenever the
daily updated data drop is available. This Google Drive endpoint is a
\strong{README} document in portable document format (PDF) which contains a
privacy and confidentiality statement, technical notes with regard to the
latest data, technical notes with regard to previous (archive data) and two
shortened URLs - one linking to the Google Drive folder that contains all
the latest officially released datasets, and the other linking to the
datasets released previously (archives). Of these, the first shortened URL
linking to the Google Drive folder containing the latest officially released
datasets is different for everyday and can only be obtained through the
\strong{README} document released for a specific day.
}
\details{
The function \code{\link[=datadrop_latest_id]{datadrop_latest_id()}} reads that PDF file, extracts the
shortened URL for the latest official released datasets written in that file,
expands that shortened URL and then extracts the unique Google Drive ID for
the latest officially released datasets. With this Google Drive ID, other
functions can then be used to retrieve information and data from the Google
Drive specified by this ID.

The DoH Data Drop archives, on the other hand, is distributed via a fixed
shortened URL (\href{https://bit.ly/DataDropArchives}{bit.ly/DataDropArchives})
which links/points to a Google Drive folder containing the previous Data Drop
releases.

The function \code{\link[=datadrop_archive_id]{datadrop_archive_id()}} expands that shortened URL and then
extracts the unique Google Drive ID for the DoH Data Drop archives folder.
With this Google Drive ID, other functions can then be used to retrieve
information and data from the Google Drive specified by this ID.
}
\examples{
datadrop_latest_id()
datadrop_archive_id(.date = "2020-11-01")
datadrop_id()
datadrop_id(version = "archive", .date = "2020-11-01")

}
\author{
Ernest Guevarra
}
