% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{datadrop_ls}
\alias{datadrop_ls}
\title{List files and folders within a specified DoH Data Drop Google Drive folder}
\usage{
datadrop_ls(id)
}
\arguments{
\item{id}{A 33 character string for the Google Drive ID of the latest
officially released DoH Data Drop. This can be obtained using the
\code{\link[=datadrop_latest_id]{datadrop_latest_id()}} function or by manually expanding the shortened URL
provided in \url{bit.ly/DataDropPH} and extracting the
Google Drive ID.}
}
\value{
A tibble containing information on the various files and folders in
the specified Google Drive for DoH Data Drop.
}
\description{
List files and folders within a specified DoH Data Drop Google Drive folder
}
\examples{
id <- datadrop_latest_id()

datadrop_ls(id = id)

}
\author{
Ernest Guevarra
}
